Make serde and num-traits dependencies optional
num-traits is only used for num_traits::Zero, which isn't used in webrender,
probably only in servo. For webrender it can be turned off to reduce
dependencies, besides - all it does is to construct an Au(0)

serde is also not needed in webrender, only for debugging, so that can
be turned off, too.

For backwards compatibility these features are turned ON by default,
in order to not break any application in the upgrade process.
Also added testing for the features in isolation to the .travis.yml